admin/pyproject.toml
[project]
name = "shop_admin"
version = "0.1.0"
description = "This service is responsible for the administration of the shop."
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
  "django-ory-auth @ git+https://github.com/micro-company/django_ory_auth@master",
  'asgiref==3.9.2',
  'attrs==25.4.0',
  'certifi==2026.1.4',
  'charset-normalizer==3.4.4',
  'deprecated==1.2.18',
  'django==6.0.1',
  'django-unfold>=0.49.0',
  'django-crispy-forms>=2.3',
  'django-debug-toolbar==6.2.0',
  'django-environ==0.12.0',
  'django-health-check==3.20.8',
  'django-prometheus @ git+https://github.com/django-commons/django-prometheus@master',
  'django-request-logging==0.7.5',
  'djangorestframework==3.16.1',
  'markdown',
  'django-filter',
  'drf-spectacular==0.28.0',
  'drf-yasg==1.21.14',
  'googleapis-common-protos==1.72.0',
  'grpcio==1.76.0',
  'idna==3.11',
  'importlib-metadata==8.7.1', # TODO: depencies of opentelemetry-api
  'inflection==0.5.1',
  'json-log-formatter==1.1.1',
  'jsonschema==4.26.0',
  'jsonschema-specifications==2025.9.1',
  'opentelemetry-api==1.39.1',
  'opentelemetry-exporter-otlp==1.39.1',
  'opentelemetry-exporter-otlp-proto-common==1.39.1',
  'opentelemetry-exporter-otlp-proto-grpc==1.39.1',
  'opentelemetry-exporter-otlp-proto-http==1.39.1',
  'opentelemetry-instrumentation==0.60b1',
  'opentelemetry-instrumentation-dbapi==0.60b1',
  'opentelemetry-instrumentation-django==0.60b1',
  'opentelemetry-instrumentation-logging==0.60b1',
  'opentelemetry-instrumentation-requests==0.60b1',
  'opentelemetry-instrumentation-sqlite3==0.60b1',
  'opentelemetry-instrumentation-wsgi==0.60b1',
  'opentelemetry-proto==1.39.1',
  'opentelemetry-sdk==1.39.1',
  'opentelemetry-semantic-conventions==0.60b1',
  'opentelemetry-util-http==0.60b1',
  'packaging==26.0',
  'pip==25.3',
  'prometheus-client==0.24.1',
  'protobuf==6.33.5', # TODO: depencies of opentelemetry-proto
  'pytz==2025.2',
  'pyyaml==6.0.3',
  'referencing==0.37.0',
  'requests==2.32.5',
  'rpds-py==0.27.1',
  'setuptools==80.10.2',
  'sqlparse==0.5.5',
  'typing-extensions==4.15.0',
  'uritemplate==4.2.0',
  'urllib3==2.6.3',
  'uv==0.9.28',
  'wrapt>=1.10,<2',
  'zipp==3.23.0',
  'psycopg2-binary==2.9.11',
  'redis',
  'django-redis',
  "mimesis>=19.1.0",
  "django-map-widgets>=0.5.1",
  "gunicorn>=23.0.0",
  "whitenoise>=6.8.0",
  "grpcio-tools>=1.76.0",
  "shortlink-org-shop-oms-grpc-python==1.76.0.2.20260201221610+71e3c8b3c462",
  "shortlink-org-shop-delivery-rpc-grpc-python>=1.76.0.2.20260201222625",
  "django-money>=3.6.0",
  "django-import-export>=4.4.0",
  "django-simple-history>=3.11.0",
]

[project.optional-dependencies]
test = [
  "pytest>=8.0",
  "pytest-django>=4.5",
  "pytest-cov>=4.0",
  "pytest-mock>=3.12",
]

[build-system]
requires = ["setuptools~=80.9"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "admin.settings"
python_files = ["test_*.py"]
pythonpath = ["src"]
addopts = "-ra -q --cov=src --cov-report=term-missing"
testpaths = ["tests"]

[tool.ruff]
unsafe-fixes = true
line-length = 120
extend-exclude = [
  "src/domain/goods/migrations/*",
  "src/domain/offices/migrations/*",
  "tests/*",
]
exclude = [".venv", "*_pb2.py", "*_pb2_grpc.py", "*_pb2.pyi"]

[tool.ruff.lint]
select = [
  # TODO: enable all checks
  #  "ALL",
  "E", "W", # pycodestyle
  "F",      # pyflakes
  "UP",     # pyupgrade
  "D",      # pydocstyle
  "N",      # pylint
  "PL",     # pylint
  "RUF",    # ruff
  "TID",    # tidy
  "I",      # isort
]

[tool.ruff.lint.per-file-ignores]
# Django manage.py pattern uses import inside try/except
"src/manage.py" = ["PLC0415"]
"src/made.py" = ["PLC0415"]
"src/migration.py" = ["PLC0415"]

[tool.ruff.lint.isort]
no-lines-before = ["standard-library", "local-folder"]
known-third-party = []
known-local-folder = ["src"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ty]
# ty type checker configuration
# https://docs.astral.sh/ty/

[tool.ty.environment]
python-version = "3.13"

[tool.ty.rules]
# Enable strict type checking
unresolved-import = "error"

[[tool.uv.index]]
url = "https://buf.build/gen/python"
